LPD-92610 Extract reusable database schema export service
